requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Celery and Background Tasks
celery==5.3.6
redis==5.0.1

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-celery==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.5.1
playwright==1.40.0

# Code Quality
black==24.1.1
ruff==0.1.14
mypy==1.8.0
types-python-dateutil==2.8.19.20240106
types-pytz==2024.1.0.20240203

# Development
ipython==8.20.0
